requirements.txt
pyaudio
faster-whisper
piper-tts
requests
json
google-generativeai
